requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
nvidia-nccl-cu12==2.27.3
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
toml==0.10.2
tzdata==2025.2
xgboost==3.0.2
